+++ aerich/ddl/mysql/__init__.py
@@ -1,10 +1,6 @@
-from typing import Type
-
-from tortoise import Model
 from tortoise.backends.mysql.schema_generator import MySQLSchemaGenerator
 
 from aerich.ddl import BaseDDL
-from aerich.exceptions import NotSupportError
 
 
 class MysqlDDL(BaseDDL):
@@ -28,9 +24,4 @@ class MysqlDDL(BaseDDL):
     _DROP_FK_TEMPLATE = "ALTER TABLE `{table_name}` DROP FOREIGN KEY `{fk_name}`"
     _M2M_TABLE_TEMPLATE = "CREATE TABLE `{table_name}` (`{backward_key}` {backward_type} NOT NULL REFERENCES `{backward_table}` (`{backward_field}`) ON DELETE CASCADE,`{forward_key}` {forward_type} NOT NULL REFERENCES `{forward_table}` (`{forward_field}`) ON DELETE CASCADE){extra}{comment}"
     _MODIFY_COLUMN_TEMPLATE = "ALTER TABLE `{table_name}` MODIFY COLUMN {column}"
-
-    def alter_column_null(self, model: "Type[Model]", field_describe: dict):
-        raise NotSupportError("Alter column null is unsupported in MySQL.")
-
-    def set_comment(self, model: "Type[Model]", field_describe: dict):
-        raise NotSupportError("Alter column comment is unsupported in MySQL.")
+    _RENAME_TABLE_TEMPLATE = "ALTER TABLE `{old_table_name}` RENAME TO `{new_table_name}`"