+++ aerich/inspectdb.py
@@ -15,6 +15,7 @@ class InspectDb:
         "LONGTEXT": "    {field} = fields.TextField({null}{default}{comment})",
         "TEXT": "    {field} = fields.TextField({null}{default}{comment})",
         "DATETIME": "    {field} = fields.DatetimeField({null}{default}{comment})",
+        "FLOAT": "    {field} = fields.FloatField({null}{default}{comment})",
     }
 
     def __init__(self, conn: BaseDBAsyncClient, tables: Optional[List[str]] = None):